Fixed performance issues and "lonely blue dot"
- Resizing the window was extremely slow and stuttering. The
application performance and memory usage also seemed to spike. And the
workspace opened with that blue dot.

All of these issues rooted from the inclusion of painter code in the
overridden paintEvent() in SpriteMainWindow.

Fixed both of those issues by making a custom helper method
updateWorkspace() to paint and removed that painter code from
overridden paintEvent() method.

REASON: Since it override the SpriteMainWIndow’s paintEvent(), whenever
the GUI was initialized, the setupUi() calls update() which calls
paintEvent(). And since this paintEvent() was overridden to do our
workspace stuff, it not only painted that dot but also kept the painter
running and kept reassigning the pixmap dropping the performance
significantly.

From now on, call updateWorkspace() and not update() whenever the
“workspace” needs to be updated.

Commented out the paintEvent() for now since cant find any use. Let’s
keep it until final phases.