GUI arrangements
- Workspace now resembles more like the other tools where the canvas is
centered on the workspace (and remains same size even if the window
zoomed). The canvas only zooms when you use zoom controls or touch
gestures (may need to do that).

But this layout will make it easier to draw on the workspace since the
pixmap will be the size of passed resolution (centered on the
workspace) and will only scale when zooming.
- Other UI tweaks.