Proper size frames in frames pane, preview
- Now the frames are displayed in proper size WITHOUT SCALING
irrespective of the original size (small 25x25 pixmaps or large 500x500
pixmaps, all fill the set square size). Thus, original pixmap details
is preserved.

This is achieved by setting the min and max size of the frame to be
added before adding it to the scroll pane.

If ever need to add a frame in our code in the future, use this:

Frame* s =currentSprite.getFrames().last();
    s->setMinimumSize(160,160);
    s->setMaximumSize(160,160);
    s->setSizePolicy(QSizePolicy::Minimum, QSizePolicy::Minimum);
    s->setScaledContents(true);