Basic Zooming, bug fixes...
- Zoom in and out menu item zooms in/out the current frame by 2x resp
(shortcut keys: (command + =) and (cmd + -) ).
- Pressing (command + 0) resets the frame to original size.
- Once you switch to other frame or add a frame, resets to the original
size. THIS WAS A DESIGN DECISION AND IS NOT A BUG.
- Before saving, all the frames are made sure to be scaled back to the
original size. Previewing also makes sure all the frames are original
size and no frame retains the scaled size.

- Fixed a bug where the duplicate frame was duplicating the last frame
in the sprite and not the current selected frame.

NOTE: I do have code written to preserve zoom across frames and such,
but I am still deciding to pursue it or trash it depending on the
performance. Right now, its subject to more rigorous testing.